Fix up incorrect size of <svg:image> with a SVG image without intrinsic size
This patch fixes up incorrect sizing of SVG Image with referred SVG image
without intrinsic dimensions such as not viewport and so on.
So when referred image is SVG, this patch checks whether dimensions of
referred image is intrinsic or not.
If this image is not intrinsic, set intrinsic size of <svg:images>.

Bug: 1045932
Change-Id: If979b3a9a7980bde9e2bee38ba2863cf223b4f0b